PrescriptionPatientList/GetItemListforSalesPage/NewGetItemListforSalesPage

+++ Libraries/HIMS.Services/Inventory/ItemListForSalesPageDTO.cs
@@ -5,7 +5,7 @@
         public long? StoreId { get; set; }
         public long ItemId { get; set; }
         public string ItemName { get; set; }
-        public float BalanceQty { get; set; }
+        public double BalanceQty { get; set; }
         public decimal? LandedRate { get; set; }
         public decimal? UnitMRP { get; set; }
         public decimal? PurchaseRate { get; set; }
@@ -14,5 +14,22 @@
         public float? IGSTPer { get; set; }
         public string? UOM { get; set; }
     }
+    //public class NewItemListForSalesPageDTO
+    //{
+    //    public long? StoreId { get; set; }
+    //    public long ItemId { get; set; }
+    //    public string ItemName { get; set; }
+    //    public double BalanceQty { get; set; }
+    //    public decimal? LandedRate { get; set; }
+    //    public decimal? UnitMRP { get; set; }
+    //    public decimal? PurchaseRate { get; set; }
+    //    public float? CGSTPer { get; set; }
+    //    public float? SGSTPer { get; set; }
+    //    public float? IGSTPer { get; set; }
+    //    public string? UOM { get; set; }
+    //    //public decimal? HmrpStrip { get; set; }
+    //    //public decimal? HmrpUnitPrice { get; set; }
+
+    //}
    
 }
